refactor(web): remove cached-api barrel file
