--- apps/web/src/utils/cached-api.ts
@@ -1,24 +1,0 @@
-export {
-  getCarComparisonData,
-  getCarHistoricalTrends,
-  getCarMarketShareData,
-  getCarMonthsList,
-  getCarsComparison,
-  getCarsData,
-  getCarTopPerformersData,
-  getLatestCarMonth,
-  getMakesList,
-  getTopMakes,
-  getTopTypes,
-} from "@web/lib/data/cars";
-export {
-  getCOEComparisonData,
-  getCOEMarketShareData,
-  getCOEMonthsList,
-  getCOEResults,
-  getCOEResultsByPeriod,
-  getCOETrendData,
-  getLatestCOEMonth,
-  getLatestCOEResults,
-  getPQPData,
-} from "@web/lib/data/coe";
